feat(payment): implement ZaloPay payment integration
- Implement payment strategy pattern with PaymentStrategy interface and factory
- Add ZaloPay payment integration using ZaloPayPaymentStrategy
- Create CashPaymentStrategy for traditional payment method
- Restructure utility classes into appropriate packages
- Add payment status and method enums
- Update invoice and order entities with payment-related fields
- Add database migrations for payment fields
- Enhance cart functionality with new CartVm and CartItemVm
- Update UI components to support new payment flow
- Implement QR code display service for ZaloPay payments
- Add invoice export service

+++ src/main/java/com/nhom4/nhtsstore/ui/page/product/ProductEditPanel.java
@@ -20,6 +20,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Date;
@@ -161,8 +162,8 @@ public class ProductEditPanel extends JPanel implements RoutablePanel {
             try {
                 Product updatedProduct = product != null ? product : new Product();
                 updatedProduct.setName(nameField.getText());
-                updatedProduct.setSalePrice(((Number) salePriceField.getValue()).doubleValue());
-                updatedProduct.setPurchasePrice(((Number) purchasePriceField.getValue()).doubleValue());
+                updatedProduct.setSalePrice(BigDecimal.valueOf(((Number) salePriceField.getValue()).doubleValue()));
+                updatedProduct.setPurchasePrice(BigDecimal.valueOf(((Number) purchasePriceField.getValue()).doubleValue()));
                 updatedProduct.setManufacturer(manufacturerField.getText());
                 updatedProduct.setManufactureDate(manufactureDatePicker.getDate());
                 updatedProduct.setExpiryDate((Date) expiryDatePicker.getDate());
